requirements.txt
backports-datetime-fromisoformat==1.0.0
rauth==0.7.3
requests==2.21.0
xmltodict==0.12.0
